Deterministic tally and remove inter tally caching.
The order of the constraints matters for the speed of the tallying
algorithm. Fix the order with a simple heuristic, picking smaller terms
first.

This allows most slow tests to finish under the timeout limit.

Remove type hash-consing and caching of results across tally calls for
now. For some reason, things get slower instead of getting faster across
calls. Investigating this issue separately.