ignore impossible branches (#100)
This PR improve on PR #98 by (almost) fixing #36 und #56 in a more general way. PR #98 can be closed.

Further, this PR makes constraint simplification much easier because we no longer need to apply the result of tally during constraint simplification. Instead, we only use tally for checking satisifiability of a constraint to ignore certain branches.

There is still a problem, see the comment on #36 which is being worked on.